requirements.txt
pylint
python-daemon
